Experimenting w/ETS
Tried out porting darklaunch to ETS to see how the code would look. Looks like
an overall win to me as the amount of code is less and darklaunch config info
is stored in ETS memory space instead of the darklaunch process'. This should
allow us to store lots of flag entries without much performance impact.

Pushing this to a branch for review and feedback.