Resource now properly responds to POST requests
Removed old tests for the Webmachine resource.  The new changes broke a lot of the tests, and since mocking the resource is a pain, we're going to hold off on automated tests until we have a CI environment in place.  At that point, we can just test the REST endpoints directly.