More gracefully handle config file quirks.
If a config path is given, but there's no actual file at that location, but the location is writable, darklaunch will now write an empty JSON configuration (i.e., "{}") to that path, allowing the application to start up.  A warning is logged if this happens, as this could potentially be surprising to unsuspecting users.

If the given config path is not writable (whether a file is there or not), darklaunch will not start.

This fix addresses a situation that arose from some cookbook testing in which the configuration file was accidentally not written to disk, causing darklaunch to shutdown and restart itself in an infinite loop.