xdarklaunch: ignore old-style darklaunch, false (not error) on missing
xdarklaunch now only looks at the header value when searching for
values. The on disk old-style darklaunch JSON file is ignored. If no
header is present, xdarklaunch will return a false value for all
is_enabled queries.

This patch also changes the behavior of queries for missing darklaunch
flags. Previously, it was an error to ask is_enabled for a flag not
present in the header. The initial motivation for this strict behavior
was to prevent critical darklaunch config errors during migration. For
non-migration work, it is more convenient and less error prone in
terms of deploying config to have the default for missing values be
'false'.

A future patch will add an is_enabled_strict/2 function that provides
the error behavior. This will make the strict checking opt-in for use
with critical migration switching.